ai_core/requirements.txt
openai
pdfplumber
python-docx
pandas
openpyxl
fastapi
uvicorn
python-multipart
tabulate